backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
passlib[bcrypt]==1.7.4
PyJWT==2.8.0
python-multipart==0.0.9
pydantic==2.8.2
google-generativeai==0.7.2
pypdf==4.2.0
python-docx==1.1.2
motor==3.3.2
pymongo==4.6.1
beanie==1.24.0

